+++ src/main/java/com/beust/jcommander/JCommander.java
@@ -1036,7 +1036,7 @@ public class JCommander {
                                 + "key" + parameter.getAssignment()
                                 + "value");
             }
-            if (def != null) {
+            if (def != null && !pd.isHelp()) {
                 String displayedDef = Strings.isStringEmpty(def.toString())
                         ? "<empty string>"
                         : def.toString();